LPS-80124 (announcements-uad) Use BaseModelUADDisplay for UADDisplay base implementation
