+++ modules/apps/collaboration/announcements/announcements-uad/src/main/java/com/liferay/announcements/uad/display/BaseAnnouncementsEntryUADDisplay.java
@@ -1,0 +1,57 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.announcements.uad.display;
+
+import com.liferay.announcements.kernel.model.AnnouncementsEntry;
+import com.liferay.announcements.uad.constants.AnnouncementsUADConstants;
+import com.liferay.user.associated.data.display.BaseModelUADDisplay;
+
+import java.util.Locale;
+
+/**
+ * Provides the base implementation for the announcements entry UAD display.
+ *
+ * <p>
+ * This implementation exists only as a container for the default methods
+ * generated by ServiceBuilder. All custom methods should be put in
+ * {@link AnnouncementsEntryUADDisplay}.
+ * </p>
+ *
+ * @author William Newbury
+ */
+public abstract class BaseAnnouncementsEntryUADDisplay
+	extends BaseModelUADDisplay<AnnouncementsEntry> {
+
+	public String getApplicationName() {
+		return AnnouncementsUADConstants.APPLICATION_NAME;
+	}
+
+	public String[] getDisplayFieldNames() {
+		return new String[] {"title", "content"};
+	}
+
+	public String getKey() {
+		return AnnouncementsUADConstants.CLASS_NAME_ANNOUNCEMENTS_ENTRY;
+	}
+
+	public String getTypeDescription() {
+		return "Announcements posted by the user";
+	}
+
+	public String getTypeName(Locale locale) {
+		return "AnnouncementsEntry";
+	}
+
+}
